requirements.txt
# Core AI/ML Dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# PDF Processing
PyMuPDF>=1.23.0

# Web Framework
streamlit>=1.28.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Semantic Chunking
nltk>=3.8.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Utility Libraries
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# Development & Testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Logging and Monitoring
structlog>=23.1.0